improved mobile nav, and fixed missing icons
Improved mobile nav by making it center correctly, have the "x" to close
it align correctly on mobile safari, make the overlay tint to cover the
entire screen (even while scrolling), and keep the menu fixed to the
same place on the screen.
Also, the menu close icon was missing, and the reddit icon displayed on
the login page was using the old class still.